Refactor code for better readability
This refactors the package for improved readability. StealCordova
becomes a type and the methods are part of its prototype.

API stays the same.
